# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "SuperMakers"
        }
    ],
    "paths": {
...
        "/api/default-answers": {
            "get": {
                "tags": [
                    "Default Answers"
                ],
                "summary": "Get a list of default answers",
                "description": "Returns a list of all default answers.",
                "operationId": "810ed4e058a798795773d6f5ca05a2a6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Default Answers"
                ],
                "summary": "Create a new default answer",
                "description": "Creates a new default answer",
                "operationId": "d56eb89bd46eece9eb5848e4d0fe3472",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_id",
                                    "answer"
                                ],
                                "properties": {
                                    "question_id": {
                                        "description": "ID of the question this answer is associated with",
                                        "type": "integer"
                                    },
                                    "answer": {
                                        "description": "Text of the default answer",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Default answer created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/default-answers/{id}": {
            "get": {
                "tags": [
                    "Default Answers"
                ],
                "summary": "Get a default answer by ID",
                "description": "Returns a single default answer",
                "operationId": "26b8213da0e77e886b08227a505e61c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the default answer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Default answer not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Default Answers"
                ],
                "summary": "Update a default answer",
                "description": "Updates an existing default answer",
                "operationId": "b387177ae09f856328f6670eded33f14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the default answer to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "question_id": {
                                        "description": "ID of the question this answer is associated with",
                                        "type": "integer"
                                    },
                                    "answer": {
                                        "description": "Text of the default answer",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Default answer updated successfully"
                    },
                    "404": {
                        "description": "Default answer not found"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Default Answers"
                ],
                "summary": "Delete a default answer",
                "description": "Deletes a default answer by ID",
                "operationId": "ee449ec0a98b858d04d6d88a8505c944",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the default answer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Default answer deleted successfully"
                    },
                    "404": {
                        "description": "Default answer not found"
                    }
                }
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Get a list of favorites",
                "description": "Returns a list of all favorite items for the authenticated user.",
                "operationId": "bc433e786ab5a0bec81c53c578e0b6aa",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a favorite item",
                "description": "Adds a new item to the user's favorites list",
                "operationId": "d426a8b38354581cb4082ac4bc62e22f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_id"
                                ],
                                "properties": {
                                    "item_id": {
                                        "description": "ID of the item to favorite",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Favorite item created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/favorites/{id}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove a favorite item",
                "description": "Removes an item from the user's favorites list",
                "operationId": "397d5ad2363742284a8d87c5dc366148",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the favorite item to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Favorite item deleted"
                    },
                    "404": {
                        "description": "Favorite item not found"
                    }
                }
            }
        },
...
    },
...
}